feat: name change to jenesis (#45)
* feat: name change to jenesis

* feat: changed to jenesis-templates

* feat: import corrections

* feat: test corrections

* feat: lint fix
